Add modified pipeline error detection script
